+++ package.json
@@ -16,7 +16,6 @@
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-progress": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-toast": "^1.2.4",
     "@radix-ui/react-tooltip": "^1.1.4",
     "@tanstack/react-table": "^8.20.6",
     "class-variance-authority": "^0.7.1",
