+++ src/app/page.tsx
@@ -1,9 +1,8 @@
-import LoginForm from "./loginpage/page";
 
 export default function LoginPage() {
   return (
     <main className='flex items-center justify-center md:h-screen'>
-      <LoginForm />
+     <p>Landing Page...</p>
     </main>
   );
 }
